Added wrapper around avatar object
We needed a way to let a loader know that the avatar failed to
upload because the storage quota was exceeded. The postAvatar
function is called from a static async task that is used by a loader,
so we didn't want to pass in an activity and a callback like we do
in other places.