Gives Ints the gift of decimal points.
Fixes MBL-4988

This is really self explanatory so have a double joke.
A guy walks into a bar and asks for 1.014 root beers.
The bartender says, “I’ll have to charge you extra, that’s a
root beer
So the guy says, “In that case, better make it a double.”